February 21 2018 -- I have things sorta set-up and working, but I am thinking now I might want to have done things differently. Am going to keep working on what I have here though, and if things do get too confusing then I will start again with a different arrangement of what does what. Right now the issue I am having is a have objects called 'managers' that represent types of objects that will exist in the game. These managers are in-charge of managing all aspects of their respective object types, such as creation, deletion, and tracking. This also includes drawing the objects.
Seems like instead I would want a way for all the objects to be drawn by a Render class instead of having rendering distributed to each object type. Oh well. It's the generator class that does the actual rendering at the end of the day, but there is a large amount of variation in how the generator will be used to draw.

i'm rambling here, but yeah i'm still trying to work things out to work out properly.